RiftGuard 0.2.3
Added
- Animation on Card Collector
- Animation on Upgrading
- Show Cost when in Shop
- Adding button to hide and unhide Card Collector

+++ RiftGuard/Assets/Script/Base/ResourceUI.cs
@@ -20,7 +20,7 @@ public class ResourceUI : MonoBehaviour
 
     public void ShowMoney()
     {
-        moneyUI.text = "$ " + this.GetComponent<BaseScript>().money.ToString();
+        moneyUI.text = "$ " + ((int)this.GetComponent<BaseScript>().money).ToString();
     }
 
     public void ShowHealth()
